refactor(web): migrate jsonb-query-generator to feature folder
