+++ apps/web/src/tools/jsonb-query-generator/index.ts
@@ -1,0 +1,5 @@
+import type { ToolModule } from "@/tools/types";
+
+import { JsonbQueryGenerator } from "./ui";
+
+export const tool: ToolModule = { id: "jsonb-query-generator", Component: JsonbQueryGenerator };
